Transfer func coeffs converted to long double for qf_poly compatibility

+++ filtersintez.h
@@ -69,8 +69,8 @@ private:
     float Fc;
     int Nfil;
 
-    QVector<float> coeffB;
-    QVector<float> coeffC;
+    QVector<long double> coeffB;
+    QVector<long double> coeffC;
 
     Filter::FType ftyp;
